experiments/mvp_adk_agents/mvp_adk_agent_without_ca/requirements.txt
google-adk[bigquery]>=1.3.0
google-auth
